backend/requirements.txt
django
djangorestframework
environs
pillow